Added separate apps for website and accounts
