+++ accounts/views.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+
+def login(request):
+    return render(request, 'login.html')
+
+def register(request):
+    return render(request, 'register.html')
