+++ swiftrail/urls.py
@@ -14,17 +14,10 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
-from . import views
+from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', views.index, name='index'),
-    path('login/', views.login, name='login'),
-    path('register/', views.register, name='register'),
-    path('pnr-status/', views.pnr_status, name='pnr-status'),
-    path('book-ticket/', views.book_ticket, name='book-ticket'),
-    path('train-enquiry/', views.train_enquiry, name='train-enquiry'),
-    path('train-search/', views.train_search, name='train-search'),
-    path('train-schedule/', views.train_schedule, name='train-schedule'),
+    path('',include('website.urls')),
+    path('accounts/', include('accounts.urls'),)
 ]
