+++ website/views.py
@@ -3,12 +3,6 @@ from django.shortcuts import render
 def index(request):
     return render(request, 'index.html')
 
-def login(request):
-    return render(request, 'login.html')
-
-def register(request):
-    return render(request, 'register.html')
-
 def pnr_status(request):
     return render(request, 'pnr-status.html')
